Enable account creation  - support foreign tutor creation  - fix means of contact bug  - extends VoCardRequestService to support documents creation  - wire validatio step  - fix i18n

+++ BackOfficeNG/modules/frontoffice/controllers/RequestCreationController.groovy
@@ -275,31 +275,38 @@ class RequestCreationController {
                             (listFieldToken[0]): listWrapper[listFieldToken[0]].get(Integer.valueOf(listFieldToken[1]))
                            ]
             }
-            else if (submitAction[1] == 'addRole' && params."owner-${submitAction[3]}" != '') {
-                def individualParam = targetAsMap(submitAction[3])
+            else if (submitAction[1] == 'addRole' && params."owner-${submitAction[3]}" != '' && params."role-${submitAction[3]}" != '') {
+                def roleParam = targetAsMap(submitAction[3])
                 def ownerIndex = Integer.valueOf(params."owner-${submitAction[3]}")  
-                def owner = objectToBind.individuals.adults[ownerIndex]
+                def owner = objectToBind.individuals."${roleParam.ownerType}"[ownerIndex]
                 def role = RoleType.forString(params."role-${submitAction[3]}")
                 def individual = null
-                if (individualParam.individualIndex != null) {
-                    individual = objectToBind.individuals."${individualParam.individualType}"[Integer.valueOf("${individualParam.individualIndex}")]
+                if (roleParam.individualIndex != null) {
+                    individual = objectToBind.individuals."${roleParam.individualType}"[Integer.valueOf("${roleParam.individualIndex}")]
                 }
                 if (role == RoleType.HOME_FOLDER_RESPONSIBLE) {
                     objectToBind.individuals.adults.eachWithIndex { adult, index ->
                         homeFolderService.removeRole(adult, null, role)
                     }
+                    objectToBind.requester = owner
                 }
                 homeFolderService.addRole(owner, individual, role)
             }
             else if (submitAction[1] == 'removeRole') {
-                def individualParam = targetAsMap(submitAction[3])
-                def owner = objectToBind.individuals.adults[Integer.valueOf("${individualParam.ownerIndex}")]
-                def role = RoleType.forString(individualParam.role)
+                def roleParam = targetAsMap(submitAction[3])
+                def owner = objectToBind.individuals."${roleParam.ownerType}"[Integer.valueOf("${roleParam.ownerIndex}")]
+                def role = RoleType.forString(roleParam.role)
                 def individual = null
-                if (individualParam.individualIndex != null) {
-                    individual = objectToBind.individuals."${individualParam.individualType}"[Integer.valueOf("${individualParam.individualIndex}")]
+                if (roleParam.individualIndex != null) {
+                    individual = objectToBind.individuals."${roleParam.individualType}"[Integer.valueOf("${roleParam.individualIndex}")]
                 }
                 homeFolderService.removeRole(owner, individual, role)
+            }
+            else if (submitAction[1] == 'tutorsEdit') {
+                session[uuidString].isTutorsEdit = true
+            }
+            else if (submitAction[1] == 'tutorsEndEdit') {
+                session[uuidString].isTutorsEdit = false
             }
             // standard save action
             else {
@@ -326,7 +333,9 @@ class RequestCreationController {
                     cRequest.setMeansOfContact(meansOfContactService.getMeansOfContactByType(moce))
                     
                     def docs = documentAdaptorService.deserializeDocuments(newDocuments, uuidString)
-                    if (SecurityContext.currentEcitizen == null) 
+                    if (requestTypeInfo.label == 'VO Card Request')
+                        requestService.create(cRequest, objectToBind.individuals.adults, objectToBind.individuals.children, objectToBind.requester.adress, docs)
+                    else if (SecurityContext.currentEcitizen == null) 
                         requestService.create(cRequest, objectToBind.requester, null, docs)
                     else if (!cRequest.draft) 
                         requestService.create(cRequest, docs)
